l6 moved AppConfig.java
