+++ lesson6/src/main/java/ru/lazarev/hibernate/service/Service.java
@@ -2,10 +2,7 @@ package ru.lazarev.hibernate.service;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.stereotype.Component;
-import ru.lazarev.hibernate.AppConfig;
 import ru.lazarev.hibernate.dao.CustomerDAO;
 import ru.lazarev.hibernate.dao.OrderDAO;
 import ru.lazarev.hibernate.dao.ProductDAO;
